// migration-newip-to-ts/src/components/view/news/newsData.interface.ts
export interface NewsData {
    author: string;
    content: string;
    description: string;
    publishedAt: string;
    source: {
        id: string;
        name: string;
    };
    title: string;
    url: string;
    urlToImage: string;
}

export interface AllNewsData {
    status: string;
    totalResults?: number;
    articles?: [];
}